+++ examples/ws2812/arduino.go
@@ -1,4 +1,4 @@
-// +build !digispark,!arduino
+// +build arduino
 
 package main
 
@@ -6,4 +6,4 @@ import "machine"
 
 // Replace neo in the code below to match the pin
 // that you are using if different.
-var neo machine.Pin = machine.NEOPIXELS
+var neo = machine.D2
